requirements.txt

requests==2.26.0
icalendar==4.0.7
python-dotenv==0.19.1
twilio==7.2.0
pytz==2021.1
